Move most things into the canister_method module
