+++ src/compiler/typescript_to_rust/azle_generate/src/canister_method/heartbeat/rust.rs
@@ -1,8 +1,7 @@
-use crate::{generators::canister_methods, ts_ast::AzleFnDecl};
+use crate::{canister_method::rust, ts_ast::AzleFnDecl};
 
 pub fn generate(heartbeat_fn_decl: &AzleFnDecl) -> proc_macro2::TokenStream {
-    let call_to_heartbeat_js_function =
-        canister_methods::generate_call_to_js_function(heartbeat_fn_decl);
+    let call_to_heartbeat_js_function = rust::generate_call_to_js_function(heartbeat_fn_decl);
 
     let function_name = heartbeat_fn_decl.get_function_name();
