+++ src/compiler/typescript_to_rust/azle_generate/src/canister_method/init/rust.rs
@@ -1,15 +1,12 @@
-use crate::{generators::canister_methods, ts_ast::AzleFnDecl};
+use crate::{canister_method::rust, ts_ast::AzleFnDecl};
 
-pub fn generate_init_method_body(
-    init_fn_decl_option: Option<&AzleFnDecl>,
-) -> proc_macro2::TokenStream {
+pub fn generate(init_fn_decl_option: Option<&AzleFnDecl>) -> proc_macro2::TokenStream {
     let function_name = match init_fn_decl_option {
         Some(init_fn_decl) => init_fn_decl.get_function_name(),
         None => "DOES_NOT_EXIST".to_string(),
     };
 
-    let call_to_init_js_function =
-        canister_methods::maybe_generate_call_to_js_function(&init_fn_decl_option);
+    let call_to_init_js_function = rust::maybe_generate_call_to_js_function(&init_fn_decl_option);
 
     quote::quote! {
         BOA_CONTEXT_REF_CELL.with(|box_context_ref_cell| {
