+++ src/compiler/typescript_to_rust/azle_generate/src/canister_method/mod.rs
@@ -1,10 +1,22 @@
-use cdk_framework::act::CanisterMethods;
+use cdk_framework::act::{
+    node::{
+        canister_method::{CanisterMethodType, QueryMethod, QueryOrUpdateDefinition, UpdateMethod},
+        CandidType, CanisterMethod, Param,
+    },
+    CanisterMethods,
+};
 
-use crate::ts_ast::TsAst;
+use crate::ts_ast::{azle_type::AzleType, AzleFnDecl, TsAst};
 
 pub use errors::ParseError;
 
-mod heartbeat_method;
+mod heartbeat;
+mod init;
+mod inspect_message;
+mod post_upgrade;
+mod pre_upgrade;
+mod query_and_update;
+mod rust;
 
 pub mod errors;
 
@@ -12,7 +24,7 @@ impl TsAst {
     pub fn build_canister_methods(&self) -> CanisterMethods {
         let heartbeat_method = self.build_heartbeat_method();
         let init_method = Some(self.build_init_method());
-        let inspect_message_method = self.build_inspect_method();
+        let inspect_message_method = self.build_inspect_message_method();
         let post_upgrade_method = Some(self.build_post_upgrade_method());
         let pre_upgrade_method = Some(self.build_pre_upgrade_method());
         let query_methods = self.build_query_methods();
@@ -27,5 +39,78 @@ impl TsAst {
             query_methods,
             update_methods,
         }
+    }
+}
+
+impl<'a> AzleFnDecl<'a> {
+    pub fn build_canister_method_node(
+        &self,
+        canister_method_type: &CanisterMethodType,
+    ) -> CanisterMethod {
+        let body = query_and_update::generate_query_and_update_body(&self);
+        let is_manual = self.is_manual();
+        let is_async = self.is_promise();
+        let name = self.get_function_name();
+        let params = self.build_params();
+        let return_type = self.build_return_type();
+
+        match canister_method_type {
+            CanisterMethodType::Query => CanisterMethod::Query(QueryMethod {
+                definition: QueryOrUpdateDefinition::new(
+                    is_async,
+                    is_manual,
+                    None,
+                    name,
+                    params,
+                    return_type,
+                    body,
+                    "azle".to_string(),
+                ),
+            }),
+            CanisterMethodType::Update => CanisterMethod::Update(UpdateMethod {
+                definition: QueryOrUpdateDefinition::new(
+                    is_async,
+                    is_manual,
+                    None,
+                    name,
+                    params,
+                    return_type,
+                    body,
+                    "azle".to_string(),
+                ),
+            }),
+            _ => panic!("TODO: YOU SHOULDN'T BE TRYING TO PARSE NON QUERY/UPDATE METHODS HERE!"),
+        }
+    }
+
+    pub fn build_params(&self) -> Vec<Param> {
+        let names = self.get_param_name_idents();
+        let types = self.build_param_types();
+        names
+            .iter()
+            .enumerate()
+            .map(|(i, name)| Param {
+                name: name.clone().to_string(),
+                candid_type: types[i].clone(),
+            })
+            .collect()
+    }
+
+    fn build_return_type(&self) -> CandidType {
+        let return_ts_type = self.get_return_ts_type();
+        let return_azle_type = AzleType::from_ts_type(return_ts_type.clone(), self.source_map);
+        return_azle_type.to_data_type()
+    }
+
+    // TODO why is this separated from get_name. It would be much simpler
+    // imho to get the names and the params all in the same pass
+    fn build_param_types(&self) -> Vec<CandidType> {
+        self.get_param_ts_types()
+            .iter()
+            .map(|ts_type| {
+                let azle_type = AzleType::from_ts_type(ts_type.clone().clone(), self.source_map);
+                azle_type.to_data_type()
+            })
+            .collect()
     }
 }
