+++ src/compiler/typescript_to_rust/azle_generate/src/canister_method/pre_upgrade/rust.rs
@@ -1,10 +1,8 @@
-use crate::{generators::canister_methods, ts_ast::AzleFnDecl};
+use crate::{canister_method::rust, ts_ast::AzleFnDecl};
 
-pub fn generate_pre_upgrade_method_body(
-    pre_upgrade_fn_decl_option: Option<&AzleFnDecl>,
-) -> proc_macro2::TokenStream {
+pub fn generate(pre_upgrade_fn_decl_option: Option<&AzleFnDecl>) -> proc_macro2::TokenStream {
     let call_to_pre_upgrade_js_function =
-        canister_methods::maybe_generate_call_to_js_function(&pre_upgrade_fn_decl_option);
+        rust::maybe_generate_call_to_js_function(&pre_upgrade_fn_decl_option);
 
     let function_name = match pre_upgrade_fn_decl_option {
         Some(pre_upgrade_fn_decl) => pre_upgrade_fn_decl.get_function_name(),
