+++ src/compiler/typescript_to_rust/azle_generate/src/canister_method/query_and_update/update.rs
@@ -1,8 +1,6 @@
 use cdk_framework::act::node::canister_method::{CanisterMethod, CanisterMethodType, UpdateMethod};
 
-use crate::ts_ast::azle_program::HelperMethods;
-
-use super::TsAst;
+use crate::{ts_ast::azle_program::HelperMethods, TsAst};
 
 impl TsAst {
     pub fn build_update_methods(&self) -> Vec<UpdateMethod> {
