+++ src/compiler/typescript_to_rust/azle_generate/src/ts_ast/azle_fn_decl/mod.rs
@@ -6,7 +6,6 @@ use swc_ecma_ast::{BindingIdent, FnDecl, Pat, TsEntityName, TsType};
 
 use crate::{canister_method_annotation::CanisterMethodAnnotation, ts_ast::GetName};
 
-pub mod canister_method_builder;
 pub mod errors;
 pub mod get_dependencies;
