+++ src/compiler/typescript_to_rust/azle_generate/src/ts_ast/ts_ast/mod.rs
@@ -6,12 +6,9 @@ use swc_ecma_parser::{lexer::Lexer, Parser, StringInput, Syntax, TsConfig};
 use crate::ts_ast::{source_map::SourceMapped, AzleProgram};
 
 mod build_candid_types;
-mod build_query_methods;
-mod build_update_methods;
 mod build_vm_value_conversion;
 mod cross_canister_calls;
 mod stable_b_tree_map;
-mod system_canister_method_builder;
 mod to_act;
 
 pub struct TsAst {
